# swagger.yaml
swagger: "2.0"
info:
    version: '1.0'
    title: 'SE2-HerokuSiffredi'
    description: ''
    license:
        name: MIT
        url: https://github.com/zanna-37/se2-HerokuSiffredi
basePath: /
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
    /users:
        get:
            summary: Returns a list of all users
            tags:
            - users
            responses:
                200:
                    description: Array of User objects
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/User'
        post:
            summary: Create a new User
            tags:
            - users
            parameters:
            -   in: body
                name: user
                description: The User that needs to be created
                schema:
                    type: object
                    properties:
                        studentNumber:
                            type: integer
                            minimum: 0
                        name:
                            type: string
                        surname:
                            type: string
                    required:
                    - studentNumber
                    - name
                    - surname
            responses:
                201:
                    description: User created
                    schema:
                        type: object
                        properties:
                            id:
                                type: integer
                                minimum: 0
                                description: User ID
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'
        put:
            summary: Update some tasks
            tags:
            - users
            parameters:
            -   in: body
                name: tasks
                description: Batch of users to be updated
                schema:
                    type: array
                    items:
                        $ref: '#/definitions/User'
            responses:
                204:
                    description: Updated Users
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'
        delete:
            summary: Delete a batch of tasks
            tags:
            - users
            parameters:
            -   in: body
                name: taskIDs
                schema:
                    type: array
                    items:
                        type: integer
                        minimum: 0
            responses:
                204:
                    description: Deleted Users
                404:
                    description: Users not found
                    schema:
                        $ref: '#/definitions/Error'
    /users/{id}:
        get:
            summary: Returns a specific user
            tags:
            - users
            parameters:
            -   in: path
                name: id
                description: ID of user
                required: true
                type: integer
                minimum: 0
            responses:
                200:
                    description: User object with given ID
                    schema:
                        $ref: '#/definitions/User'
                400:
                    description: Specified ID is not valid
                404:
                    description: No User for that ID
        post:
            summary: Not supported
            deprecated: true
            tags:
            - users
            parameters:
            -   in: path
                name: id
                required: true
                type: integer
                minimum: 0
            responses:
                405:
                    description: Method not allowed
                    schema:
                        $ref: '#/definitions/Error'
        put:
            summary: Updates a specific user
            tags:
            - users
            parameters:
            -   in: path
                name: id
                description: ID of user
                required: true
                type: integer
                minimum: 0
            -   in: body
                name: user
                schema:
                    $ref: '#/definitions/User'
            responses:
                204:
                    description: Updated User
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'
                404:
                    description: No User for that ID
                    schema:
                        $ref: '#/definitions/Error'
        delete:
            summary: Delete specific user
            tags:
            - users
            parameters:
            -   in: path
                name: id
                description: ID of user
                required: true
                type: integer
                minimum: 0
            responses:
                204:
                    description: Deleted User
                404:
                    description: No User for that ID
                    schema:
                        $ref: '#/definitions/Error'
...
    /submissions:
        get:
            summary: Not supported
            deprecated: true
            tags:
            - submissions
            responses:
                405:
                    description: Method not allowed
                    schema:
                        $ref: '#/definitions/Error'

        post:
            summary: New Submission
            tags:
            - submissions
            parameters:
            -   in: body
                name: submission
                description: The submission that must be created
                schema:
                    type: object
                    properties:
                        examID:
                            type: integer
                            minimum: 0
                        userID:
                            type: integer
                            minimum: 0
                        assignedTaskID:
                            type: integer
                            minimum: 0
                        userAnswer:
                            type: string
                        finalCorrectionID:
                            type: integer
                            minimum: 0
                    required:
                    - examID
                    - userID
                    - assignedTaskID
                    - userAnswer
            responses:
                201:
                    description: Submission created
                    schema:
                        type: object
                        properties:
                            id:
                                type: integer
                                minimum: 0
                                description: Submission ID
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'

        put:
            summary: Update some submissions
            tags:
            - submissions
            parameters:
            -   in: body
                name: submissions
                description: Batch of submissions to be updated
                schema:
                    type: array
                    items:
                        $ref: '#/definitions/Submission'
            responses:
                204:
                    description: Updated Submission
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'

        delete:
            summary: Delete some submissions
            tags:
            - submissions
            responses:
                204:
                    description: Deleted Submission
                404:
                    description: No Submission for that ID
                    schema:
                        $ref: '#/definitions/Error'
    /submissions/{id}:
        get:
            summary: Returns a specific submission
            tags:
            - submissions
            parameters:
            -   in: path
                name: id
                description: ID of submission
                required: true
                type: integer
                minimum: 0
            responses:
                200:
                    description: submission with given ID
                    schema:
                        $ref: '#/definitions/Submission'
                400:
                    description: Specified ID is not valid
                404:
                    description: No submission for that ID
        post:
            summary: Not supported
            deprecated: true
            tags:
            - submissions
            parameters:
            -   in: path
                name: id
                required: true
                type: integer
                minimum: 0
            responses:
                405:
                    description: Method not allowed
                    schema:
                        $ref: '#/definitions/Error'
        put:
            summary: Updates a specific submission
            tags:
            - submissions
            parameters:
            -   in: path
                name: id
                description: ID of submission
                required: true
                type: integer
                minimum: 0
            -   in: body
                name: submission
                schema:
                    $ref: '#/definitions/Submission'
            responses:
                204:
                    description: Updated Submission
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'
                404:
                    description: No Submission for that ID
                    schema:
                        $ref: '#/definitions/Error'
        delete:
            summary: Delete specific Submission
            tags:
            - submissions
            parameters:
            -   in: path
                name: id
                description: ID of submission
                required: true
                type: integer
                minimum: 0
            responses:
                204:
                    description: Deleted submission
                404:
                    description: No submission for that ID
                    schema:
                        $ref: '#/definitions/Error'

        patch:
            summary: Correction submitted by the professor
            tags:
            - submissions
            parameters:
            -   in: path
                name: id
                description: ID of submission
                required: true
                type: integer
                minimum: 0
            -   in: body
                name: correction
                schema:
                    properties:
                        submissionID:
                            type: integer
                            minimum: 0
                        finalCorrectionID:
                            type: integer
                            minimum: 0
                    required:
                    - submissionID
                    - finalCorrectionID
            responses:
                204:
                    description: Sumbission corrected
                400:
                    description: Bad request
                    schema:
                        $ref: '#/definitions/Error'
                404:
                    description: No submission for that ID
                    schema:
                        $ref: '#/definitions/Error'
definitions:
...
    User:
        title: User
        type: object
        properties:
            id:
                type: integer
                minimum: 0
            studentNumber:
                type: integer
                minimum: 0
            name:
                type: string
            surname:
                type: string
            average:
                type: number
        required:
        - id
        - studentNumber
...
    Submission:
        title: Submission
        type: object
        properties:
            submissionID:
                type: integer
                minimum: 0
            examID:
                type: integer
                minimum: 0
            userID:
                type: integer
                minimum: 0
            assignedTaskID:
                type: integer
                minimum: 0
            userAnswer:
                type: string
            finalCorrectionID:
                type: integer
                minimum: 0
        required:
        - submissionID
        - examID
        - userID
        - assignedTaskID
        - userAnswer
    Error:
        title: Error
        type: object
        properties:
            code:
                type: integer
            message:
                type: string
        required:
        - code
        - message